+++ lib/assets_handlers/generators_data.dart
@@ -6,6 +6,7 @@ import 'package:realm_roller/pages/guilds/guild_generation/guild_generation_page
 import 'package:realm_roller/pages/kingdoms/kingdom_generation/kingdom_generation_page.dart';
 import 'package:realm_roller/pages/landscapes/landscape_generation/landscape_generation_page.dart';
 import 'package:realm_roller/pages/locations/locations_generation/location_generation_page.dart';
+import 'package:realm_roller/pages/names/names_generation/companion_names_generator.dart';
 import 'package:realm_roller/pages/names/names_generation/npc_names_generation_page.dart';
 import 'package:realm_roller/pages/names/names_generation/settlement_name_generation.dart';
 import 'package:realm_roller/pages/npcs/npc_generation/npc_generation_page.dart';
@@ -74,6 +75,11 @@ const Map<String, GeneratorData> generatorsData = {
     title: "Companions",
     icon: Icons.pets,
     generatorPage: CompanionGenerationPage(),
+  ),
+  "Companion Names": GeneratorData(
+    title: "Companion Names",
+    icon: Icons.cruelty_free,
+    generatorPage: CompanionNameGenerationPage(),
   ),
   "Emblems": GeneratorData(
     title: "Emblems",